requirements.txt
langchain
langchain-openai
python-dotenv
langchain-groq
langchain-community
bs4
unstructured
pypdf
wikipedia
#langchian-chroma
chromadb
openai
langserve
uvicorn
#pydantic==1.10.13
langgraph
llama-index
openai